Rename trapFocus to trap-focus, add dnt-enabled
